Refactor UIKit extensions into separate files.
