+++ Chat App/CustomComponents/String/String+Base64.swift
@@ -1,0 +1,21 @@
+//
+//  String+Base64.swift
+//  Chat App
+//
+//  Created by Amiel Jireh Cordova on 12/28/20.
+//  Copyright © 2020 Cordova, Jireh. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension String {
+    func base64Encoded() -> String? {
+        return data(using: .utf8)?.base64EncodedString()
+    }
+
+    func base64Decoded() -> String? {
+        guard let data = Data(base64Encoded: self) else { return nil }
+        return String(data: data, encoding: .utf8)
+    }
+}
